+++ src/Akeneo/Pim/Enrichment/Component/Product/Normalizer/Standard/Product/ProductValuesNormalizer.php
@@ -14,9 +14,15 @@ use Symfony\Component\Serializer\SerializerAwareTrait;
  * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductValuesNormalizer implements NormalizerInterface, SerializerAwareInterface
+class ProductValuesNormalizer implements NormalizerInterface
 {
-    use SerializerAwareTrait;
+    /** @var NormalizerInterface */
+    private $normalizer;
+
+    public function __construct(NormalizerInterface $normalizer)
+    {
+        $this->normalizer = $normalizer;
+    }
 
     /**
      * {@inheritdoc}
@@ -26,7 +32,7 @@ class ProductValuesNormalizer implements NormalizerInterface, SerializerAwareInt
         $result = [];
 
         foreach ($data as $value) {
-            $normalizedValue = $this->serializer->normalize($value, $format, $context);
+            $normalizedValue = $this->normalizer->normalize($value, $format, $context);
             $result[$value->getAttributeCode()][] = $normalizedValue;
         }