+++ src/Akeneo/Pim/Enrichment/Component/Product/Query/GetConnectorProducts.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Component\Product\Query;
+
+use Akeneo\Pim\Enrichment\Component\Product\Connector\ReadModel\ConnectorProductList;
+
+/**
+ * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+interface GetConnectorProducts
+{
+    /**
+     * Ideally, we should not pass the PQB but a Product Query agnostic of the storage.
+     * It would be much easier fake it.
+     */
+    public function fromProductQueryBuilder(
+        ProductQueryBuilderInterface $productQueryBuilder,
+        ?array $attributesToFilterOn,
+        ?string $channelToFilterOn,
+        ?array $localesToFilterOn
+    ): ConnectorProductList;
+}