+++ tests/back/Pim/Enrichment/EndToEnd/Product/Product/ExternalApi/AbstractProductTestCase.php
@@ -40,6 +40,8 @@ abstract class AbstractProductTestCase extends ApiTestCase
         $this->get('pim_catalog.saver.product')->save($product);
 
         $this->get('akeneo_elasticsearch.client.product')->refreshIndex();
+        $this->get('akeneo_elasticsearch.client.product_and_product_model')->refreshIndex();
+
 
         return $product;
     }
@@ -68,6 +70,8 @@ abstract class AbstractProductTestCase extends ApiTestCase
         $this->get('pim_catalog.saver.product')->save($product);
 
         $this->get('akeneo_elasticsearch.client.product')->refreshIndex();
+        $this->get('akeneo_elasticsearch.client.product_and_product_model')->refreshIndex();
+
 
         return $product;
     }
@@ -97,6 +101,7 @@ abstract class AbstractProductTestCase extends ApiTestCase
         $this->get('pim_catalog.saver.product_model')->save($productModel);
 
         $this->get('akeneo_elasticsearch.client.product_model')->refreshIndex();
+        $this->get('akeneo_elasticsearch.client.product_and_product_model')->refreshIndex();
 
         return $productModel;
     }