+++ tests/back/Pim/Enrichment/Specification/Bundle/Elasticsearch/IdentifierResultCursorSpec.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Enrichment\Bundle\Storage\ElasticsearchAndSql\ProductGrid;
+namespace Specification\Akeneo\Pim\Enrichment\Bundle\Elasticsearch;
 
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\IdentifierResult;
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
 use PhpSpec\ObjectBehavior;
 
-class ProductAndProductModelIdentifierCursorSpec extends ObjectBehavior
+class IdentifierResultCursorSpec extends ObjectBehavior
 {
     function let(IdentifierResult $identifierResult1, IdentifierResult $identifierResult2)
     {
         $this->beConstructedWith([$identifierResult1, $identifierResult2], 42);
     }
 
-    function it_is_a_Cursor()
+    function it_is_a_cursor()
     {
         $this->shouldImplement(CursorInterface::class);
     }